+++ ydb/query/transaction.py
@@ -290,14 +290,12 @@ class BaseQueryTxContext(base.CallbackHandler, Generic[DriverT]):
     @overload
     def _begin_call(
         self: "BaseQueryTxContext[SyncDriver]", settings: Optional[BaseRequestSettings]
-    ) -> "BaseQueryTxContext[SyncDriver]":
-        ...
+    ) -> "BaseQueryTxContext[SyncDriver]": ...
 
     @overload
     def _begin_call(
         self: "BaseQueryTxContext[AsyncDriver]", settings: Optional[BaseRequestSettings]
-    ) -> Awaitable["BaseQueryTxContext[AsyncDriver]"]:
-        ...
+    ) -> Awaitable["BaseQueryTxContext[AsyncDriver]"]: ...
 
     def _begin_call(
         self, settings: Optional[BaseRequestSettings]
@@ -319,14 +317,12 @@ class BaseQueryTxContext(base.CallbackHandler, Generic[DriverT]):
     @overload
     def _commit_call(
         self: "BaseQueryTxContext[SyncDriver]", settings: Optional[BaseRequestSettings]
-    ) -> "BaseQueryTxContext[SyncDriver]":
-        ...
+    ) -> "BaseQueryTxContext[SyncDriver]": ...
 
     @overload
     def _commit_call(
         self: "BaseQueryTxContext[AsyncDriver]", settings: Optional[BaseRequestSettings]
-    ) -> Awaitable["BaseQueryTxContext[AsyncDriver]"]:
-        ...
+    ) -> Awaitable["BaseQueryTxContext[AsyncDriver]"]: ...
 
     def _commit_call(
         self, settings: Optional[BaseRequestSettings]
@@ -349,14 +345,12 @@ class BaseQueryTxContext(base.CallbackHandler, Generic[DriverT]):
     @overload
     def _rollback_call(
         self: "BaseQueryTxContext[SyncDriver]", settings: Optional[BaseRequestSettings]
-    ) -> "BaseQueryTxContext[SyncDriver]":
-        ...
+    ) -> "BaseQueryTxContext[SyncDriver]": ...
 
     @overload
     def _rollback_call(
         self: "BaseQueryTxContext[AsyncDriver]", settings: Optional[BaseRequestSettings]
-    ) -> Awaitable["BaseQueryTxContext[AsyncDriver]"]:
-        ...
+    ) -> Awaitable["BaseQueryTxContext[AsyncDriver]"]: ...
 
     def _rollback_call(
         self, settings: Optional[BaseRequestSettings]
@@ -390,8 +384,7 @@ class BaseQueryTxContext(base.CallbackHandler, Generic[DriverT]):
         arrow_format_settings: Optional[base.ArrowFormatSettings],
         concurrent_result_sets: Optional[bool],
         settings: Optional[BaseRequestSettings],
-    ) -> Iterable[_apis.ydb_query.ExecuteQueryResponsePart]:
-        ...
+    ) -> Iterable[_apis.ydb_query.ExecuteQueryResponsePart]: ...
 
     @overload
     def _execute_call(
@@ -407,8 +400,7 @@ class BaseQueryTxContext(base.CallbackHandler, Generic[DriverT]):
         arrow_format_settings: Optional[base.ArrowFormatSettings],
         concurrent_result_sets: Optional[bool],
         settings: Optional[BaseRequestSettings],
-    ) -> Awaitable[Iterable[_apis.ydb_query.ExecuteQueryResponsePart]]:
-        ...
+    ) -> Awaitable[Iterable[_apis.ydb_query.ExecuteQueryResponsePart]]: ...
 
     def _execute_call(
         self,